move weighed_average() and substitute_nonexistent_values() to new flowsbyclean.py
